internal/cli/cmd: fix operator image autodetection
[ cherry-picked from cilium/cilium-cli repository ]

Commit a855451c7e7a ("internal/cli/cmd: allow specifying agent and
operator image via CLI parameters") broke the operator image
autodetection for AWS and Azure which leads to `operator-generic` being
deployed rather than `operator-{aws.azure}`.

This is due to the fact that `utils.BuildImagePath` will only set the
provided cloud-specific image path if the OperatorImage parameter is
empty. Fix this by making the default CLI option value empty, so the
autodetection will kick in as long as the option is not specified.

Fixes: a855451c7e7a ("internal/cli/cmd: allow specifying agent and operator image via CLI parameters")
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>