connectivity: tests/world - run all curls from all client Pods
[ cherry-picked from cilium/cilium-cli repository ]

Before, a random client Pod was selected, leading to a degree of randomness
that cannot easily be accounted for in verdict overrides. In reality, the
check was often executed from the same client Pod, defeating the point of the
test.

This patch executes all three curls from all available client Pods.

Signed-off-by: Timo Beckers <timo@isovalent.com>