connectivity: maintain a single feed of flows per Action
[ cherry-picked from cilium/cilium-cli repository ]

There is no documentation around why this is the way it is, but this looks
like it's keeping multiple copies of the same flow feed. There is only one
source that already contains all packets in both directions, so it's not
clear why this is. Source/destination of a packet is universal, it doesn't
depend on where it's being observed from.

There are still a lot of improvements to be made here, but this implementation
will correctly open a flow follower against the existing Hubble Relay client
and use it to populate the Action's 'flows' field. This is to be replaced with
a fully-streaming matcher later, which would allow for canceling long-running
commands that are only expected to generate drops, to avoid having to wait for
them to time out.

The flow feed includes all packets to and from the Pod executing the test,
which should be all we need to perform verification. All currently-known Hubble
errors are handled and unknown errors terminate the suite.

Some more refactoring and simplification is highly needed in the matcher itself,
since the 'offset' system is no longer required (and fragile to begin with).
Merging matcher results should also be redundant. A full verdict can be emitted
in a single pass, since the flow feed is only open-ended towards the future,
meaning no past packets can be backfilled into the buffer. Again, a complex
implementation that proved hard to understand without any existing comments,
docs or tests.

This also removes some more boilerplate from Scenarios and streamlines the harness
a bit. Ideally, we would omit FlowParameters entirely and only have actual
test logic in the scenario. Validating flows is the job of the harness itself,
test code shouldn't be expected to have to worry about this. (it just gets
copy-pasta'd anyway...)

Signed-off-by: Timo Beckers <timo@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>