workflows: use latest stable CLI in post-test information gathering
[ cherry-picked from cilium/cilium-cli repository ]

The post-test information gathering steps were previously self-contained
and became split between the actual workflow step and in-cluster jobs
logs in #212.

In parallel, `sysdump` became a subcommand of the CLI and we switched to
using it in post-test information gathering in #405.

However, we explicitly do not want to run PR-built `cilium` in the
context of a `pull_request_target`-triggered privileged workflow, as it
would allow an attacker to leak the repository's secrets.

Our proposal is to use the latest stable version of the CLI in the post-
test information gathering step. This is acceptable as the purpose of
this step is retrieving information, not testing the actual `cilium`
commands run in the step -- this is done as part of the in-cluster jobs.

Should a workflow developer actually need to check that a PR editing
`sysdump` or any other `cilium` command used in this step works as
intended with the new changes, they can always switch to `pull_request`
testing or edit the in-cluster scripts with additional testing.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>