install: Disable BPF masquerading by default
[ cherry-picked from cilium/cilium-cli repository ]

In commit bf2c1f99 ("install: Disable kube-proxy-replacement by
default") we switched kube-proxy-replacement to disabled by default. We
didn't change the default setting for BPF masquerading, even though it
requires kube-proxy-replacement to be enabled. As a result, default
installation result in a warning when BPF masquerading is forcefully
disabled.

This commit disables BPF masquerading by default to fix it. We could
enable BPF masquerading when kube-proxy-replacement is enabled by the
user, but it seems more confusing than anything.

Fixes: bf2c1f99 ("install: Disable kube-proxy-replacement by default")
Signed-off-by: Paul Chaignon <paul@cilium.io>