kpr: Improve k8s API server endpoint detection to handle Kind
[ cherry-picked from cilium/cilium-cli repository ]

When kubeproxy is missing on the cluster, Cilium CLI automatically
sets KPR mode to strict and finds API server endpoint from kubernetes
config. This doesn't work in the Kind environment since the API server
endpoint on the config can only accessed through port-forwarding on the
host.

To make it work, we need to resolve the real API server endpoint which
can be accessed by cilium-agent. One way to achieve that is relying on
`kubernetes` Endpoints object which shows us the endpoints of
`kubernetes` service (k8s API service).

This commit enhances existing k8s API server endpoint detection to
handle Kind case. When we detect the Kind environment, it lookups the
`kubernetes` Endpoints, extract first IP + Port combination, and use
them as apiServerHost and apiServerPort.

Before the change:

```
ℹ️  helm template --namespace kube-system cilium cilium/cilium --version 1.12.2 --set bpf.monitorAggregation=none,cluster.id=1,cluster.name=cluster1,debug.enabled=true,encryption.nodeEncryption=false,ipam.mode=kubernetes,k8sServiceHost=127.0.0.1,k8sServicePort=40421,kubeProxyReplacement=strict,operator.replicas=1,serviceAccounts.cilium.name=cilium,serviceAccounts.operator.name=cilium-operator,tunnel=vxlan
```

After the change:

```
ℹ️  helm template --namespace kube-system cilium cilium/cilium --version 1.12.2 --set bpf.monitorAggregation=none,cluster.id=1,cluster.name=cluster1,debug.enabled=true,encryption.nodeEncryption=false,ipam.mode=kubernetes,k8sServiceHost=172.18.0.2,k8sServicePort=6443,kubeProxyReplacement=strict,operator.replicas=1,serviceAccounts.cilium.name=cilium,serviceAccounts.operator.name=cilium-operator,tunnel=vxlan
```

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>