Adds a test case for L7 policy with TLS
[ cherry-picked from cilium/cilium-cli repository ]

This adds a test to check if HTTPS traffic can get inspected in a CNP.
To enable this it adds new helper functions to provision secrets,
certificates and a public CA bundle into the test setup.
It also adds the functionality to insert a file into a pod,
this is done for having the certificate chain checked in curl.

Signed-off-by: Maartje Eyskens <maartje.eyskens@isovalent.com>