connectivity: Make detected Cilium Agent version available for feature detection
[ cherry-picked from cilium/cilium-cli repository ]

Make detected Cilium Agent version available for feature detection and
use it instead of trying to detect Cilium version from the agent image
reference. This helps avoid failures like this when the image tag is not
a semantic version (in this case image tag is "test"):

  connectivity test failed: unable to determine if KNP feature is enabled: failed to parse Cilium version test to semver

Cilium agent version was already detected, this PR moves the detection a
bit earlier to make it available for feature detection.

Fixes: #1470

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>