connectivity: Ability to add custom annotations to test deployment
[ cherry-picked from cilium/cilium-cli repository ]

This commit adds the ability to add custom namespace and pod annotations
to the `cilium connectivity test` deployment. The two flags are
`--deployment-pod-annotations` and `--namespace-annotations`.

For the pod annotations, we accept a JSON map which contains the
deployment name as the key, and the annotations as a string-to-string
map. For the namespace annotation, we simply expect a string-to-string
map.

We could have used Viper's `StringToString` map for the namespace
annotations, but not for the pod annotations, since there we would need
a "`StringToStringToString`" map. Therefore, to remain consistent
between the two flags, both flags exclusively JSON syntax.

If the deployment name in the annotations map ends with a `*` symbol,
then we perform a longest prefix match.

The flags are currently hidden, since we are not fully commited to this
command-line syntax yet and it might still change in the future.

Example:

```
$ cilium connectivity test \
    --namespace-annotations='{"foo":"bar"}' \
    --deployment-pod-annotations='{"client":{"baz":"qux"},"echo-*":{"quux":"corge"}}'
```

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>