k8s/copy: check file size
[ cherry-picked from cilium/cilium-cli repository ]

I've encountered an issue in a cluster, where the tetragon bugtool files
were not complete. After debugging it seems that not all contents of the
files were copied. This patch tries to handle this problem by getting
the size of the file before the copy. If it is succesful in doing so,
then it checks whether we have read everything and if not it returns an
error.

There is already support for handling partial reads in the code, so
after this patch I am able to read the full size of the files.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>