connectivity test: check for deleted cilium agent pod in health probe
[ cherry-picked from cilium/cilium-cli repository ]

Currently, the health probe connectivity test fails forever if an
Cilium Agent Pod that existed when starting the connectivity tests no
longer exists.

The reason is that the health probes uses the list of Cilium Agent pods
that got fetched at the beginning of the connectivity test run.

Therefore, this commit adds a check whether the Pod still exists. If not,
the health probe check fails.

The underlying reason is often that a underlying K8s node has been deleted
in the meantime (since starting the tests).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>