connectivity: detect PodCIDR for both local and remote nodes
[ cherry-picked from cilium/cilium-cli repository ]

The detected PodCIDRs are later used to configure the routes
required by "from-outside" tests. Let's make sure that we
retrieve them for both local and remote nodes when clustermesh
is enabled, so that we can run these tests also in that case.

From outside tests work under the assumption that the nodes (of both
clusters) are on the same LAN, and therefore we can configure a
direct route targeting them. This holds for CI workflows using Kind.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>