clustermesh: make status command compatible with external kvstore
[ cherry-picked from cilium/cilium-cli repository ]

The clustermesh-apiserver cannot be used in combination with Cilium
running in kvstore mode when the identity allocation mode is kvstore.
In this case, remote clusters shall be configured to directly connect
to the local etcd cluster. Let's extend the `clustermesh status`
command to additionally cover this setup, ignoring the validation of
clustermesh-apiserver specific resources.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>