connectivity: don't pre-populate the ingress service map
[ cherry-picked from cilium/cilium-cli repository ]

We already retrieve the services at a later stage, and the pre-populated
map is actually incorrect, as the namespace is not configured, causing
misleading failures if not replaced. To this end, a subsequent commit
will also extend the retrieval step to make sure that the service was
actually created, and hence retrieved.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>