k8s: add method to proxy connection to container's TCP port
[ cherry-picked from cilium/cilium-cli repository ]

Extent the Client implementation introducing a new method which allows
to proxy a connection to a TCP port inside a container. This mimics the
behavior of the port-forward implementation, but it directly provides
access to the forwarded stream (through a ReadWriteCloser interface),
rather than exposing it through a local port.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>